refactor(discord): harden gateway lifecycle fsm
